# apd: tune performance by inlining function calls and manually unrolling loops