This PR contains a series of follow-on micro-optimizations that build upon the changes in https://github.com/cockroachdb/apd/pull/106.

The individual changes here focus primarily on inlining function calls on the hot paths and on manually unrolling loops where possible. For the first form of optimization, it uses `gcassert` to detect regressions. I'm still hoping to add `gcassert` to CI, but that is currently blocked on https://github.com/jordanlewis/gcassert/pull/3.

Impact on benchmark suite:
```
name                 old time/op    new time/op    delta
GDA/squareroot-10      16.0ms ± 1%    12.6ms ± 0%  -21.26%  (p=0.000 n=8+9)
GDA/abs-10             6.14µs ± 1%    4.85µs ± 0%  -21.04%  (p=0.000 n=10+8)
GDA/powersqrt-10        246ms ± 0%     195ms ± 1%  -20.49%  (p=0.000 n=9+9)
GDA/reduce-10          14.0µs ± 0%    11.3µs ± 1%  -19.00%  (p=0.000 n=8+9)
GDA/minus-10           7.83µs ± 0%    6.57µs ± 1%  -16.10%  (p=0.000 n=8+10)
GDA/rounding-10         308µs ± 0%     261µs ± 1%  -15.20%  (p=0.000 n=8+10)
GDA/remainder-10       32.8µs ± 1%    28.7µs ± 1%  -12.47%  (p=0.000 n=8+10)
GDA/divide-10           268µs ± 0%     237µs ± 1%  -11.57%  (p=0.000 n=8+10)
GDA/quantize-10        81.2µs ± 1%    72.8µs ± 1%  -10.29%  (p=0.000 n=9+10)
GDA/divideint-10       15.9µs ± 1%    14.3µs ± 0%   -9.92%  (p=0.000 n=8+8)
GDA/subtract-10        95.7µs ± 1%    86.7µs ± 1%   -9.48%  (p=0.000 n=8+9)
GDA/compare-10         29.1µs ± 1%    26.6µs ± 1%   -8.87%  (p=0.000 n=10+9)
GDA/multiply-10        56.2µs ± 1%    52.2µs ± 1%   -7.24%  (p=0.000 n=9+9)
GDA/cuberoot-apd-10    1.95ms ± 0%    1.82ms ± 0%   -6.92%  (p=0.000 n=8+9)
GDA/randoms-10         1.88ms ± 1%    1.77ms ± 0%   -5.91%  (p=0.000 n=9+9)
GDA/tointegralx-10     18.4µs ± 1%    17.4µs ± 1%   -5.57%  (p=0.000 n=9+10)
GDA/ln-10              85.3ms ± 0%    80.8ms ± 0%   -5.30%  (p=0.000 n=9+9)
GDA/tointegral-10      18.0µs ± 1%    17.1µs ± 0%   -5.08%  (p=0.000 n=9+10)
GDA/log10-10            110ms ± 0%     104ms ± 0%   -4.91%  (p=0.000 n=8+8)
GDA/plus-10            36.5µs ± 1%    34.7µs ± 0%   -4.86%  (p=0.000 n=9+9)
GDA/add-10              603µs ± 0%     576µs ± 1%   -4.49%  (p=0.000 n=9+9)
GDA/power-10            227ms ± 0%     217ms ± 0%   -4.46%  (p=0.000 n=9+9)
GDA/exp-10              125ms ± 0%     121ms ± 1%   -3.12%  (p=0.000 n=8+10)
GDA/comparetotal-10    25.6µs ± 2%    24.9µs ± 2%   -2.64%  (p=0.000 n=10+10)
GDA/base-10             117µs ± 1%     117µs ± 1%     ~     (p=0.326 n=10+8)
```

The benchmark suites are valuable to give good coverage over a diverse set of operations. However, because each "op" is actually a few thousand different arithmetic operations over varying sized values, they don't give much intuition about the absolute cost of these operations. Additionally, because they try to exercise edge cases, the average operation in these tests skews expensive.

Here's the impact of this change on a single basic addition operation (`add '1.23456789'  '1.00000000' -> '2.23456789'`):
```
name        old time/op    new time/op    delta
GDA/add-10    80.9ns ± 1%    73.6ns ± 1%  -9.10%  (p=0.000 n=10+10)
```